Number Guessing Game (Simple Go Fish)
Uses python to create a random number guessing game through the console. Makes use of enumerate and randomchoice() functions. Can be increased to include more numbers by using randint()  or randrange() to create a random number for list 2. Can be accomplished without list but utilized to increase readability of text tutorial. In addition, I plan on using list 1 for more experiments later on.