Python Console Image Converter
Made a console-based image to graph converter. This version is a basic console version of the image converter. Currently, the image converter only converts to 4 cmaps but can be customized to include several other kinds of cmaps in the future. On the initial image and new image, the user must include the image type that they would like to create as indicated in the prompt. Does not work with all compilers due to matplotlib working different ways on different compilers. Image Converter was tested using Spyder (Python 3.8). May later choose to update the text to also include capitalizations for cmaps. Wanted to show off a basic version of this code for demonstration purposes.