Python Image Converter Application
The application version of the Python Image Converter. Tkinter application enables the user to make multiple changes to the image file without auto-closing (similar to adding a loop function to the initial console code. 

Currently, the image converter only converts to 4 cmaps but can be customized to include several other kinds of cmaps in the future. On the initial image and new image, the user must include the image type that they would like to create as indicated in the prompt. Does not work with all compilers due to matplotlib working different ways on different compilers. Image Converter was tested using Spyder (Python 3.8). May later choose to update the text to also include capitalizations for cmaps. Wanted to show off a basic version of this code for demonstration purposes. Can be utilized by design school students to create a graph of their image that they can work with or to see where specific parts of the image are in 2-dimensional space.

May later create future updates to allow the user to choose whether or not the image is flipped, whether the image has a color bar and several other image quality options.