Audio Recorder 1.0
Basic Console-based audio recorder that can be used for generating .wav, .mp3, and .mp4 files. Utilizes pyaudio and wave. The user sets the file that they would like to create and the amount of seconds that they would like to record for. The Audio Recorder will overwrite the file if it has exactly the same name. The project works in Python 3.8 but not 3.9 due to issues with pyaudio in Python 3.9. If one is using the Audio Recorder in Python 3.9, set the environment to Python 3.8 for best results.