Update All Nonary Values Equal 9 Proof
Code demonstrates how all multiples of 9 become 9 after self-based addition. For instance, 18 is 1 + 8 = 18, 27 is 2 + 7 = 9, 36 is 3 + 6 = 9, etc. If we utilize recursion, we can calculate most equations in O(n^3) time complexity. Most factors of 9 can be completed within 3 recursive cycles. However, ultimate time complexity is O(x^n). Where n is the number recursive cycles and x is the length of the array. Values such as 99,999 would take more recursive cycles to calculate, but this value is highly reduced by memoization inherent in the list. More known values in the nineCalculator function will significantly reduce the time required by the system to pull values, but this is just a proof.

Note: Online compilers may cap out at the 10,000 multiples due to online memory issues.