Create Pierats (How to Remove Letters from a String)
Easy way to remove specific letters from an array. Function has time complexity O(n^2), where is n is the length of the array. The space complexity is O(n) 2 in this case is comparable to x for each subsequent check through the array. A simple way of filtering through a list if the filter function isn't available. The function is set to remove all instances of p, i, and e from the string value which is currently pirates, leaving behind the word rats, hence pierats. However, it is actually more cost efficient to skip values in the array when creating an array such that if array[iterable] = p then break, which I will display in a later project. Removing the numbers in one pass at least removes the x from the possible time complexity of the function despite being just as complicated to write (if you can call this complicated).