Create Pierats 2.0 (How to Remove Letters from a String)
Using a for loop and a try except function we can reduce the Pierats function down to true O(n) time. The true time complexity for the function is O(2n) where you only ever iterate through the list twice, once to make the list, then again to remove all items to be removed. The space complexity of the function is also O(n) where n is the length of the stored list in memory. The function can be shortened by using the filter() method and creating a function that checks for the items to be removed. I would love to see a similar type of function utilized on a dictionary to return the hidden value or message in a dictionary. Be careful as i does not stand for iterable in this function. See initial description of pierats for more details.

Overall Time Complexity O(n)
Overall Space Complexity O(n)