Day 19 Turtle Race
A simple turtle racing game using the turtle module for python.

* = Things to do if you haven't already.

*pip install turtle
*pip install random

Each turtle is notated by a different color and vertical starting position. Every turtle's speed is randomized using a random integer between 1-10. The user is prompted at the beginning of the race to choose a turtle. The race will stop once the first turtle crosses the finish line. The game is simple so if there is a tie the lower turtle on the list will be declared as winner. By changing the animal parameter according to the documentation the user can configure the project to have an assortment of animals.